// apps/api/src/interfaces/IRequest.ts
import { IEducation, IEqualOp, IExperience, ILink } from "./IUser";
import { Types } from "mongoose";
import { ApplicationStatus, JobLevel } from "./IApplication";

export interface QueriesRequest {
  page: string;
  pageSize: string;
  textSearch: string;
}

export interface SaveUserRequest {
  id?: Types.ObjectId;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  education?: IEducation[];
  experience?: IExperience[];
  equalOp?: IEqualOp;
  links?: ILink[];
}

export interface SaveApplicationRequest {
  id?: Types.ObjectId;
  title?: string;
  company?: string;
  location?: string;
  level?: JobLevel;
  jobPostUrl?: string;
  description?: string;
  datePosted?: Date;
  status?: ApplicationStatus;
  createdAt?: Date;
  userId?: Types.ObjectId;
}